Refactor to use HTMLInputElement type property 🥳
Element.setAttribute() is fine, but even better: the HTMLInputElement
interface provides special properties for manipulating an `<input>`.
The `type` property will return/set the attribute of the element.

This change uses more specific syntax, and it’s easier to read now that
I leveled up and can understand how to read the docs about this API:
https://developer.mozilla.org/en-US/docs/Web/API/HTMLInputElement