Improve split delimeter to regex for line breaks
Using an empty string as a delimiter would fail when two words are
separated by a line break, counting them as a single word.

This regex pattern will catch spaces `\s` and line breaks `\n` `\r`
with a `+` to for one or more consecutive characters and a global flag
`g` at the end to keep the regex going after finding the first match.