Improve accessibility with an ARIA live region
The live update of word and character count when the user is typing is
not apparent to people with assistive technologies like a screen reader.
This change will add `aria-live` to expose dynamic content changes.

When using Voice Over to test the functionality, we learnt how the
original nested markup wasn’t optimal, because only the last update was
annonced. Just the number for the character count, and no word count.
Refactoring to a single live region with single `textContent` provides
a consistent announcement and also helpful context for the count.