Modify script to prevent duplicate quotes
We want new quotes, duplicates are boring! `indexOf()` will return the
index an item in an array. We can use this to check if the current
quote returned from the API is present in our array of recent quotes,
because this method will return `-1` if not.

`Array.push()` will add the current quote to the end of that array.
If the length of the array has reached the limit we care to check for,
`Array.shift()` will remove a quote from the beginning of the array.