Refactor the check for duplicate quotes
The previous approach checks that quote is NOT a duplicate before doing
the things. This change checks if quote IS a duplicate and then can
immediately proceed with a skip to try again. The following statements
can then be written outside any if statement and I no longer need to
nest the second if inside the first if.

Code did exactly what it should, but refactoring is good practice! And
I like how this becomes more straight forward now. Not a big deal with
only these lines, in a more complex function I think having simpler
steps can make a huge difference in readability.