Use NYT API to display their top stories on food 🍳
Let's practice using `fetch` with a more complex API, this returns an
array of articles in the specified section and I can access a bunch of
properties (like title and abstract). The `map()` method will let me
transform the array and modify it, specifically to add markup. Today
this will get injected into the DOM with `innerHTML` but I have also
read though the security considerations on MDN.

The api-key is not commited because it should be kept secret. I can use
it locally and set up support for environment variables later.