Generate monsters with Fisher–Yates shuffle 👾
This is the start of a game! Since JavaScript doesn't have a built in
method for shuffling arrays, the recommended algorithm is Fisher–Yates
shuffle. Explained and visualised: https://bost.ocks.org/mike/shuffle/

I want to draw my own monsters, so this commit is without SVG files and
will display the image `alt` attribute for now.