Hide all monsters behind doors 🙈
Academy day 25 project is to modify the script to render doors that can
be opened. Considered if 12 doors should be put directly in the HTML.
But nope, this is a game with no progressive enhancement to talk about,
it's not like 12 closed doors is a useful fallback if no JS.

`CSS.escape()` feels like a hack I'm not supposed to have come across,
but it will let me use the value of the data attribute as a selector.
Which is here the current (clicked) item from the array of monsters.

`closest()` will traverse the element and parents to look for a match
to the provided selector. The method will return itself or closest
ancestor that matches. Using it here makes sure that both click on
button and on img element gets an event listener and fires an event.