Disable doors after game is lost 🧦
It was possible to keep opening doors after the game was completed,
which also rendered additional false messages about winning and
loosing. This change will disable remaining doors after a lost game.

But when you win, you can open the last door to verify a sock is there.
Checking that the counter is less than the length of the array makes
sure you won't get a message about losing when you open the last door.