Refactor to a niftier solution for missing IDs ⚓️
Incrementing a number as an anchor for headings without an id works,
but if the text changes — the url will change. Generating an anchor
from the heading text is a more robust solution.

`replace()` method will replace every character it matches with a
hyphen, making urls without spaces or other characters that don't work.
This regex pattern will only allow letters and numbers, `g` for global
search and `i` for case insensitive search.