Fix bug that let new input data overwrite localStorage
Data in localStorage was retrieved to repopulate the form inputs. But
the moment a user types in an input, the function to store data would
only handle new values — wiping the object clean. Now there is a check
for previously stored data also in the function to handle input events.