Improve the checkbox code
When I started coding this, I thought the checked attribute and the
state of a checkbox were the exact same thing. This caused a world of
confusion. Now I have learnt that a checkbox can be in a state of
unchecked and still have a `checked` attribute – and also the opposite.
Even though the code here doesn't have to explicitly set and remove the
checked attribute to work as intended, I think I prefer that it does.
This solution is longer than the one suggested in the courses' project
review, but I find it easier to wrap my head around when what I see in
the DOM in dev tools corresponds with the browser and the js.

Refactor also to use strings with either 'on' or 'off' to save as data,
this seems less confusing than pretending true/false is the value.
Also figured out how if else means I don't need those weird returns.